Fixes an issue that local deployment cannot switch UI (#1002)
1. For local deployment, which does not have authentication enabled,
instead of showing the user profile menu, show a "Settings" menu
instead.
2. The settings menu has a minimized command list that only supports
"switch UI" + "Privacy settings" menu items. Because "user API key" and
"Sign out" isn't applicable for user-less deployments.
3. Also added a placeholder entry point on the old UI to switch to the
new UI, instead of relying on the user profile menu to switch to the new
UI because the user profile menu does not display for the local
deployment without authentication enabled.
4. The experience for deployments with authentication doesn't change.


Preview:

![capture1](https://github.com/sematic-ai/sematic/assets/133257643/bc88f330-2d25-4458-9027-47d3bc737929)

Closes #985 
Closes #986
